--- ludaya/schema.sql
@@ -1,6 +1,0 @@
-drop table if exists entries;
-create table entries (
-  id integer primary key autoincrement,
-  title text not null,
-  'text' text not null
-);
